+++ src/usecases/delete_group_event/DeleteGroupEventDataAccessInterface.java
@@ -1,4 +1,4 @@
-package usecases.delete_personal_event;
+package usecases.delete_group_event;
 
 public interface DeleteGroupEventDataAccessInterface {
     void removeGroupEvent(String name, String eventName, String startTime, String endTime);
