+++ src/usecases/delete_group_event/DeleteGroupEventInputBoundary.java
@@ -1,4 +1,4 @@
-package usecases.delete_personal_event;
+package usecases.delete_group_event;
 
 public interface DeleteGroupEventInputBoundary {
     public void executeDelete(DeleteGroupEventInputData inputData);
