+++ src/usecases/delete_group_event/DeleteGroupEventOutputBoundary.java
@@ -1,4 +1,4 @@
-package usecases.delete_personal_event;
+package usecases.delete_group_event;
 
 public interface DeleteGroupEventOutputBoundary {
     void setPassView(DeleteGroupEventOutputData outputData);
