+++ src/usecases/delete_group_event/DeleteGroupEventOutputData.java
@@ -1,4 +1,4 @@
-package usecases.delete_personal_event;
+package usecases.delete_group_event;
 
 public class DeleteGroupEventOutputData {
     private String eventName;
